PRO-6576 fix default values for object fields in widgets and generally unify newInstance implementations
